- handling onDeleteSupportingDocumentClick() method of BorrowerPaymentFragment.kt in BorrowerPaymentViewModel.kt
#6

+++ app/src/main/java/com/rohitthebest/manageyourrenters/ui/fragments/borrower/BorrowerHomeFragment.kt
@@ -293,28 +293,36 @@ class BorrowerHomeFragment : Fragment(R.layout.fragment_borrower_home),
             && borrowerForMenus.supportingDocument != null
         ) {
 
-            if (borrowerForMenus.supportingDocument?.documentType != DocumentType.URL) {
+            showAlertDialogForDeletion(
+                requireContext(),
+                {
+                    if (borrowerForMenus.supportingDocument?.documentType != DocumentType.URL) {
 
-                if (!isInternetAvailable(requireContext())) {
+                        if (!isInternetAvailable(requireContext())) {
 
-                    showToast(
-                        requireContext(),
-                        "Network connection required to delete the document from cloud"
-                    )
-                    return
-                } else {
+                            showToast(
+                                requireContext(),
+                                getString(R.string.network_required_for_deleting_file_from_cloud)
+                            )
+                            return@showAlertDialogForDeletion
+                        } else {
 
-                    deleteFileFromFirebaseStorage(
-                        requireContext(),
-                        borrowerForMenus.supportingDocument?.documentUrl!!
-                    )
+                            deleteFileFromFirebaseStorage(
+                                requireContext(),
+                                borrowerForMenus.supportingDocument?.documentUrl!!
+                            )
+                        }
+                    }
+                    borrowerForMenus.supportingDocument = null
+                    borrowerForMenus.isSupportingDocAdded = false
+                    borrowerViewModel.updateBorrower(borrowerForMenus)
+                    showToast(requireContext(), getString(R.string.supporting_document_deleted))
+                    it.dismiss()
+                },
+                {
+                    it.dismiss()
                 }
-            }
-            borrowerForMenus.supportingDocument = null
-            borrowerForMenus.isSupportingDocAdded = false
-            borrowerViewModel.updateBorrower(borrowerForMenus)
-            showToast(requireContext(), "Supporting Document deleted")
-
+            )
         } else {
             showToast(requireContext(), getString(R.string.no_supporting_doc_added))
         }
